% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accumulated_visiting_time}
\alias{accumulated_visiting_time}
\title{Computes the accumulated visiting time (integral of stop probability) of the Markov Jump Process of the phase-type distribution standing at each vertex}
\usage{
accumulated_visiting_time(phase_type_graph, time, granularity = 0L)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{time}{The time to wait}

\item{granularity}{Optional. Number of jumps per time unit (approximation)}
}
\value{
A numeric vector of the accumulated visiting time, for each vertex
}
\description{
Computes the accumulated visiting time (integral of stop probability) of the Markov Jump Process of the phase-type distribution standing at each vertex
}
\seealso{
\code{\link[=rph]{rph()}}

\code{\link[=dph]{dph()}}

\code{\link[=pph]{pph()}}
}
