% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_edge}
\alias{add_edge}
\title{Adds an edge between two vertices in the graph}
\usage{
add_edge(
  phase_type_vertex_from,
  phase_type_vertex_to,
  weight,
  parameterized_edge_state = as.numeric(c())
)
}
\arguments{
\item{phase_type_vertex_from}{The vertex that transitions from}

\item{phase_type_vertex_to}{The vertex that transitions to}

\item{weight}{The weight of the edge, i.e. the transition rate}

\item{parameterized_edge_state}{Optional. Associate a numeric vector to an edge, for faster computations of moments when weights are changed.}
}
\description{
The graph represents transitions between states as
a weighted direction edge between two vertices.
}
\examples{
graph <- create_graph(4)
vertex_a <- find_or_create_vertex(graph, c(1,2,1,0))
vertex_b <- find_or_create_vertex(graph, c(2,0,1,0))
add_edge(vertex_a, vertex_b, 1.5)
}
\seealso{
\code{\link[=expected_waiting_time]{expected_waiting_time()}}

\code{\link[=moments]{moments()}}

\code{\link[=variance]{variance()}}

\code{\link[=covariance]{covariance()}}

\code{\link[=graph_update_weights_parameterized]{graph_update_weights_parameterized()}}
}
