% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clone_graph}
\alias{clone_graph}
\title{Clones the graph, returning a new identical graph}
\usage{
clone_graph(phase_type_graph)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\value{
A clone of the input graph
}
\description{
Clones the graph, returning a new identical graph
}
\details{
Many of the functions in this library mutate the graph. Use this to clone.
}
