% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covariance}
\alias{covariance}
\title{Computes the covariance of the phase-type distribution}
\usage{
covariance(phase_type_graph, rewards1, rewards2)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{rewards1}{First vector of rewards, which should be applied to the phase-type distribution. Must have length equal to \code{\link[=vertices_length]{vertices_length()}}}

\item{rewards2}{Second vector of rewards, which should be applied to the phase-type distribution. Must have length equal to \code{\link[=vertices_length]{vertices_length()}}}
}
\value{
The covariance of the distribution given the two rewards
}
\description{
This function invokes \code{\link[=expected_waiting_times]{expected_waiting_times()}}
twice to find the first and second moment for each of the two rewards
}
\examples{
graph <- ptdalgorithms::create_graph(4)
v1 <- ptdalgorithms::create_vertex(graph, c(1,2,3,4))
v2 <- ptdalgorithms::create_vertex(graph, c(4,0,3,3))
a <- ptdalgorithms::create_vertex(graph, c(0,0,0,0))
ptdalgorithms::add_edge(ptdalgorithms::starting_vertex(graph), v1, 1)
ptdalgorithms::add_edge(v1, v2, 4)
ptdalgorithms::add_edge(v2, a, 10)
covariance(graph, c(0,2,1,0), c(0,2,1,0)) == variance(graph, c(0,2,1,0))
covariance(graph, c(0,2,1,0), c(0,5,2,0)) # =>
  0.645
ph <- ptdalgorithms::graph_as_matrix(graph)
# This is a much faster version of
ph$IPV\%*\%solve(-ph$SIM)\%*\%diag(c(2,1))\%*\%solve(-ph$SIM)\%*\%diag(c(5,2)) \%*\% rep(1, length(ph$IPV)) + ph$IPV\%*\%solve(-ph$SIM)\%*\%diag(c(5,2))\%*\%solve(-ph$SIM)\%*\%diag(c(2,1)) \%*\% rep(1, length(ph$IPV)) - ph$IPV\%*\%solve(-ph$SIM)\%*\%diag(c(2,1)) \%*\% rep(1, length(ph$IPV)) \%*\% ph$IPV\%*\%solve(-ph$SIM)\%*\%diag(c(5,2)) \%*\% rep(1, length(ph$IPV)) # =>
  0.645
}
\seealso{
\code{\link[=expected_waiting_time]{expected_waiting_time()}}

\code{\link[=expectation]{expectation()}}

\code{\link[=variance]{variance()}}

\code{\link[=moments]{moments()}}
}
