% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_graph}
\alias{create_graph}
\title{Create a graph representing a phase-type distribution}
\usage{
create_graph(state_length)
}
\arguments{
\item{state_length}{The length of the integer vector used to represent and reference a state.}
}
\value{
Simple reference to a CPP object.
}
\description{
\code{create_graph} creates a graph representing a phase-type distribution.
This is the primary entry-point of the library.
}
\details{
There will \emph{always} be a starting vertex added to
the graph.

Notice that when the library functions are invoked on
this object, the object is \emph{mutated}, i.e. changed, which
may be surprising considering the normal behavior of R
objects.
}
\examples{
graph <- create_graph(4)
}
