% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_vertex}
\alias{create_vertex}
\title{Create a vertex matching \code{state}}
\usage{
create_vertex(phase_type_graph, state)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{state}{An integer vector of the newly added vertex. Has length as given by \code{state_length} in  \code{\link[=create_graph]{create_graph()}}}
}
\value{
The newly inserted vertex in the graph
}
\description{
Creates the vertex and adds it to
the graph object. Warning: the function \code{\link[=find_or_create_vertex]{find_or_create_vertex()}}
should be preferred. This function will \emph{not} update the lookup tree,
so \code{\link[=find_vertex]{find_vertex()}} will \emph{not} return it.
}
\seealso{
\code{\link[=find_or_create_vertex]{find_or_create_vertex()}}
}
