% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ddph}
\alias{ddph}
\title{Probability mass function of the discrete phase-type distribution}
\usage{
ddph(x, phase_type_graph)
}
\arguments{
\item{x}{Vector of the number of jumps (discrete time)}

\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\value{
A numeric vector of the density
}
\description{
Returns the density (probability mass function) at a specific number of jumps.
}
\seealso{
\code{\link[=rdph]{rdph()}}

\code{\link[=qdph]{qdph()}}
}
