% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{defect}
\alias{defect}
\title{Computes the defect, i.e. the probability of immediately transitioning to the absorbing state, of the phase-type distribution}
\usage{
defect(phase_type_graph)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\value{
A numeric value
}
\description{
Returns the defect of the distribution
}
\seealso{
\code{\link[=dph]{dph()}}

\code{\link[=qph]{qph()}}
}
