% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distribution_context}
\alias{distribution_context}
\title{Builds a probability distribution context for the phase-type distribution.}
\usage{
distribution_context(phase_type_graph, granularity = 0L)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{granularity}{Optional. Number of jumps per time unit (approximation)}
}
\value{
A probability distribution context
}
\description{
This allows the user to step through the distribution, computing e.g. the
time-inhomogeneous distribution function or the expectation of a multivariate phase-type distribution
}
