% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distribution_context_step}
\alias{distribution_context_step}
\title{Performs one jump in the probability distribution context for the discrete phase-type distribution.}
\usage{
distribution_context_step(probability_distribution_context)
}
\arguments{
\item{probability_distribution_context}{The context created by \code{\link[=distribution_context]{distribution_context()}}}
}
\description{
This allows the user to step through the distribution, computing e.g. the
time-inhomogeneous distribution function or the expectation of a multivariate phase-type distribution.
\emph{mutates} the context
}
\seealso{
\code{\link[=distribution_context]{distribution_context()}}
}
