% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dph}
\alias{dph}
\title{Probability density function of the phase-type distribution}
\usage{
dph(x, phase_type_graph, granularity = 0L)
}
\arguments{
\item{x}{Vector of the number of stopping times}

\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{granularity}{Optional. Number of jumps per time unit (approximation)}
}
\value{
A numeric vector of the density
}
\description{
Returns the density (probability density function) at a specific time. Notice that this
is an approximation
}
\seealso{
\code{\link[=rph]{rph()}}

\code{\link[=qph]{qph()}}
}
