% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dph_covariance}
\alias{dph_covariance}
\title{Computes the covariance of the discrete phase-type distribution}
\usage{
dph_covariance(phase_type_graph, rewards1, rewards2)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{rewards1}{First vector of rewards, which should be applied to the discrete phase-type distribution. Must have length equal to \code{\link[=vertices_length]{vertices_length()}}}

\item{rewards2}{Second vector of rewards, which should be applied to the discrete  phase-type distribution. Must have length equal to \code{\link[=vertices_length]{vertices_length()}}}
}
\value{
The covariance of the discrete distribution given the two rewards
}
\description{
This function invokes [ptdalgorithms::moments())]
twice to find the first and second moment for each of the two rewards

[ptdalgorithms::moments())]: R:ptdalgorithms::moments())
}
\seealso{
\code{\link[=dph_expected_visits]{dph_expected_visits()}}

\code{\link[=dph_expectation]{dph_expectation()}}

\code{\link[=dph_variance]{dph_variance()}}
}
