% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dph_distribution_context}
\alias{dph_distribution_context}
\title{Builds a probability distribution context for the discrete phase-type distribution.}
\usage{
dph_distribution_context(phase_type_graph)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\value{
A probability distribution context
}
\description{
This allows the user to step through the distribution, computing e.g. the
time-inhomogeneous distribution function or the expectation of a multivariate discrete discrete phase-type distribution
}
