% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dph_distribution_context_accumulated_visits}
\alias{dph_distribution_context_accumulated_visits}
\title{Returns the accumulated visits for the current probability distribution context for the discrete phase-type distribution.}
\usage{
dph_distribution_context_accumulated_visits(probability_distribution_context)
}
\arguments{
\item{probability_distribution_context}{The context created by \code{\link[=dph_distribution_context]{dph_distribution_context()}}}
}
\description{
This allows the user to step through the distribution, computing e.g. the
time-inhomogeneous distribution function or the expectation of a multivariate discrete phase-type distribution.
\emph{mutates} the context
}
\seealso{
\code{\link[=dph_distribution_context]{dph_distribution_context()}}
}
