% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dph_expectation}
\alias{dph_expectation}
\title{Computes the expectation (mean) of the discrete phase-type distribution}
\usage{
dph_expectation(phase_type_graph, rewards = NULL)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{rewards}{Optional rewards, which should be applied to the discrete phase-type distribution. Must have length equal to \code{\link[=vertices_length]{vertices_length()}}}
}
\value{
The expectation of the distribution
}
\description{
This function invokes \code{\link[=moments]{moments()}}
and takes the first entry (from starting vertex)
}
\seealso{
\code{\link[=dph_expected_visits]{dph_expected_visits()}}

\code{\link[=expected_waiting_time]{expected_waiting_time()}}

\code{\link[=dph_variance]{dph_variance()}}

\code{\link[=dph_covariance]{dph_covariance()}}
}
