% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dph_expected_visits}
\alias{dph_expected_visits}
\title{Computes the expected jumps (or accumulated rewards) until absorption}
\usage{
dph_expected_visits(phase_type_graph, rewards = NULL)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{rewards}{Optional rewards, which should be applied to the discrete phase-type distribution. Must have length equal to \code{\link[=vertices_length]{vertices_length()}}}
}
\value{
A numeric vector where entry \code{i} is the expected rewarded jumps starting at vertex \code{i}
}
\description{
This function can be used to compute the moments of a
disrete phase-type distribution very fast and without much
memory usage compared with the traditional matrix-based
equations.

The function takes in non-integers as rewards, but to be a \emph{strictly} valid
rewarded discrete phase-type distribution these should be integers
}
\seealso{
\code{\link[=moments]{moments()}}

\code{\link[=expectation]{expectation()}}

\code{\link[=variance]{variance()}}

\code{\link[=covariance]{covariance()}}
}
