% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dph_random_sample_stop_vertex}
\alias{dph_random_sample_stop_vertex}
\title{Samples \code{n} vertices at the stopping times of the discrete phase-type distribution}
\usage{
dph_random_sample_stop_vertex(n, phase_type_graph, jumps)
}
\arguments{
\item{n}{The number of random samples to draw}

\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{jumps}{The number of jumps before stopping}
}
\value{
A vector of numeric values, the indices of the vertices at the stopping time (jumps)
}
\description{
Samples \code{n} vertices at the stopping times of the discrete phase-type distribution
}
\seealso{
\code{\link[=rdph]{rdph()}}
}
