% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dph_stop_probability}
\alias{dph_stop_probability}
\title{Computes the probability of the Markov Chain of the discrete phase-type distribution standing at each vertex after n jumps}
\usage{
dph_stop_probability(phase_type_graph, jumps)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{jumps}{Number of jumps, discrete time}
}
\value{
A numeric vector of the stop probabilities, for each vertex
}
\description{
Computes the probability of the Markov Chain of the discrete phase-type distribution standing at each vertex after n jumps
}
\seealso{
\code{\link[=rdph]{rdph()}}

\code{\link[=ddph]{ddph()}}

\code{\link[=pdph]{pdph()}}
}
