% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dph_variance}
\alias{dph_variance}
\title{Computes the variance of the discrete phase-type distribution}
\usage{
dph_variance(phase_type_graph, rewards = NULL)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{rewards}{Optional rewards, which should be applied to the discrete phase-type distribution. Must have length equal to \code{\link[=vertices_length]{vertices_length()}}}
}
\value{
The variance of the distribution
}
\description{
This function invokes \code{\link[=moments]{moments()}}
twice to find the first and second moment
}
\seealso{
\code{\link[=dph_expected_visits]{dph_expected_visits()}}

\code{\link[=dph_covariance]{dph_covariance()}}

\code{\link[=dph_expectation]{dph_expectation()}}
}
