% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{edges}
\alias{edges}
\title{Returns the out-going edges of a vertex}
\usage{
edges(phase_type_vertex)
}
\arguments{
\item{phase_type_vertex}{The vertex to find the edges for}
}
\value{
A list of out-going edges
}
\description{
Returns a list of edges added by \code{\link[=add_edge]{add_edge()}}
}
