% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expectation}
\alias{expectation}
\title{Computes the expectation (mean) of the phase-type distribution}
\usage{
expectation(phase_type_graph, rewards = NULL)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{rewards}{Optional rewards, which should be applied to the phase-type distribution. Must have length equal to \code{\link[=vertices_length]{vertices_length()}}}
}
\value{
The expectation of the distribution
}
\description{
This function invokes \code{\link[=expected_waiting_times]{expected_waiting_times()}}
and takes the first entry (from starting vertex)
}
\examples{
graph <- phasic::create_graph(4)
v1 <- phasic::create_vertex(graph, c(1,2,3,4))
v2 <- phasic::create_vertex(graph, c(4,0,3,3))
a <- phasic::create_vertex(graph, c(0,0,0,0))
phasic::add_edge(phasic::starting_vertex(graph), v1, 1)
phasic::add_edge(v1, v2, 4)
phasic::add_edge(v2, a, 10)
phasic::expectation(graph) # =>
  0.35
phasic::expectation(graph, c(0,2,1,0)) # =>
  0.6
ph <- phasic::graph_as_matrix(graph)
# This is a much faster version of
ph$IPV\%*\%solve(-ph$SIM) \%*\% rep(1, length(ph$IPV)) # =>
  0.35
ph$IPV\%*\%solve(-ph$SIM) \%*\% diag(c(2,1))\%*\% rep(1, length(ph$IPV)) # =>
  0.35
}
\seealso{
\code{\link[=expected_waiting_time]{expected_waiting_time()}}

\code{\link[=moments]{moments()}}

\code{\link[=variance]{variance()}}

\code{\link[=covariance]{covariance()}}
}
