% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_or_create_vertex}
\alias{find_or_create_vertex}
\title{Find or create a vertex matching \code{state}}
\usage{
find_or_create_vertex(phase_type_graph, state)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{state}{An integer vector of what vertex to look for. Has length as given by \code{state_length} in  \code{\link[=create_graph]{create_graph()}}}
}
\value{
The newly found or inserted vertex in the graph
}
\description{
Finds a vertex by the \code{state} parameter. If no such
vertex exists, it creates the vertex and adds it to
the graph object instead.
}
\details{
A faster and simpler version of calling \code{\link[=find_vertex]{find_vertex()}} and  \code{\link[=create_vertex]{create_vertex()}}
}
\examples{
graph <- create_graph(4)
find_or_create_vertex(graph, c(1,2,1,0)) # Adds and returns the vertex
find_or_create_vertex(graph, c(1,2,1,0)) # Only returns the vertex
# `graph` is now changed permanently
}
