% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_vertex}
\alias{find_vertex}
\title{Finds a vertex matching \code{state}}
\usage{
find_vertex(phase_type_graph, state)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{state}{An integer vector of the newly added vertex. Has length as given by \code{state_length} in  \code{\link[=create_graph]{create_graph()}}}
}
\value{
The found vertex in the graph or NA
}
\description{
Finds a vertex matching \code{state}
}
\seealso{
\code{\link[=find_or_create_vertex]{find_or_create_vertex()}}
}
