% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{graph_as_dph_matrix}
\alias{graph_as_dph_matrix}
\title{Converts the graph-based \emph{discrete} phase-type distribution into a traditional sub-transition matrix and initial probability vector}
\usage{
graph_as_dph_matrix(phase_type_graph)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\value{
A list of the sub-transition matrix, states, and initial probability vector, and graph indices matching the matrix (1-indexed)
}
\description{
Converts the graph-based \emph{discrete} phase-type distribution into a traditional sub-transition matrix and initial probability vector
}
\details{
Used to convert to the traditional matrix-based formulation.
Has three entries: \verb{$STM} the sub-transition matrix, \verb{$IPV} the initial
probability vector, \verb{$states} the state of each vertex. Does
\emph{not} have the same order as \code{\link[=vertices]{vertices()}}
It is expected that all out-going edges have weights summing to 1 or
less, and the remaining probability is considered as a self-transition. The indices
returned are 1-based, like the input to \code{\link[=vertex_at]{vertex_at()}}
}
\examples{
graph <- ptdalgorithms::create_graph(4)
v2 <- ptdalgorithms::create_vertex(graph, c(4,0,3,3))
v1 <- ptdalgorithms::create_vertex(graph, c(1,2,3,4))
a <- ptdalgorithms::create_vertex(graph, c(0,0,0,0))
ptdalgorithms::add_edge(ptdalgorithms::starting_vertex(graph), v1, 0.5)
ptdalgorithms::add_edge(v1, v2, 0.8)
ptdalgorithms::add_edge(v2, a, 0.5)
# Note that if a defect is desired, the edge has to be explicitly
# added to an absorbing vertex
ptdalgorithms::add_edge(ptdalgorithms::starting_vertex(graph), a, 0.5)
ptdalgorithms::graph_as_dph_matrix(graph) #>
# $`states`
#         [,1] [,2] [,3] [,4]
#   [1,]    1    2    3    4
#   [2,]    4    0    3    3
# $STM
#         [,1]  [,2]
#   [1,]   0.2   0.8
#   [2,]   0.0   0.5
# $IPV
#   [1] 0.5 0
# $indices
#   [1] 3 2
}
\seealso{
\code{\link[=graph_as_matrix]{graph_as_matrix()}}
}
