% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{graph_update_weights_parameterized}
\alias{graph_update_weights_parameterized}
\title{Updates all parameterized edges of the graph by given scalars.}
\usage{
graph_update_weights_parameterized(phase_type_graph, scalars)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{scalars}{A numeric vector of multiplies for the edge states.}
}
\description{
Given a vector of scalars, computes a new weight of
the parameterized edges in the graph by a simple inner
product of the edge state vector and the scalar vector.
}
\details{
A faster and simpler version to compute new moments, when
the user wants to try multiple different weights.
}
\examples{
graph <- create_graph(4)
v1 <- find_or_create_vertex(graph, c(1,2,1,0))
v2 <- find_or_create_vertex(graph, c(2,0,1,0))
add_edge(starting_vertex(graph), v1, 5)
add_edge(v1, v2, 0, c(5,2))
edges(starting_vertex(graph))[[1]]$weight # => 5
edges(v1)[[1]]$weight # => 0
graph_update_weights_parameterized(graph, c(9,7))
edges(starting_vertex(graph))[[1]]$weight # => 5
edges(v1)[[1]]$weight # => 59
}
\seealso{
\code{\link[=expected_waiting_time]{expected_waiting_time()}}

\code{\link[=add_edge]{add_edge()}}
}
