% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_as_graph}
\alias{matrix_as_graph}
\title{Converts the the matrix-based representation into a phase-type graph}
\usage{
matrix_as_graph(IPV, SIM, rewards = NULL)
}
\arguments{
\item{IPV}{The initial probability vector (alpha)}

\item{SIM}{The sub-intensity matrix (S)}

\item{rewards}{Optional. The state/rewards of each of the vertices.}
}
\value{
A graph object
}
\description{
Converts the the matrix-based representation into a phase-type graph
}
\details{
Sometimes the user might want to use the fast graph algorithms,
but have some state-space given as a matrix. Therefore we can construct
a graph from a matrix. If desired, a discrete phase-type distribution
should just have no self-loop given. Note that the function
\code{graph_as_matrix} may reorder the vertices to make the graph represented
as strongly connected components in an acyclic manner.
}
\examples{
g <- matrix_as_graph(
    c(0.5,0.3, 0),
    matrix(c(-3, 0, 0, 2, -4, 1, 0, 1,-3), ncol=3),
    matrix(c(1,4,5,9,2,7), ncol=2)
)

graph_as_matrix(g)
}
\seealso{
\code{\link[=matrix_as_graph]{matrix_as_graph()}}
}
