% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize_dph_graph}
\alias{normalize_dph_graph}
\title{Normalizes the discrete phase-type distribution, by making self-loops go to an auxiliary vertex with reward 0 (such that all out-going weights sum to 1 for each vertex)}
\usage{
normalize_dph_graph(phase_type_graph)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\description{
Changes \code{phase_type_graph} such that all out-going weight sum to 1 for each vertex
and the self-loops go to a newly added vertex, with reward 0, that goes directly back,
as described in the paper, returns the associated new rewards. Mutates the phase-type graph permanently.
}
