% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize_graph}
\alias{normalize_graph}
\title{Normalizes the phase-type distribution, such that all out-going weights sum to 1 for each vertex}
\usage{
normalize_graph(phase_type_graph)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\description{
Changes \code{phase_type_graph} such that all out-going weight sum to 1 for each vertex
as described in the paper, returns the associated inverse rates (new rewards). Mutates the phase-type graph permanently.
}
