% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdph}
\alias{pdph}
\title{Cumulative distribution function of the discrete phase-type distribution}
\usage{
pdph(q, phase_type_graph)
}
\arguments{
\item{q}{Vector of the quantiles (jumps, discrete time)}

\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\value{
A numeric vector of the distribution function
}
\description{
Returns the density (probability mass function) at a specific number of jumps.
}
\seealso{
\code{\link[=rdph]{rdph()}}

\code{\link[=ddph]{ddph()}}
}
