% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pph}
\alias{pph}
\title{Cumulative distribution function of the phase-type distribution}
\usage{
pph(q, phase_type_graph, granularity = 0L)
}
\arguments{
\item{q}{Vector of the quantiles (stopping times)}

\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{granularity}{Optional. Number of jumps per time unit (approximation)}
}
\value{
A numeric vector of the distribution function
}
\description{
Returns the density (cumulative probability density function) at a specific time. Notice that this
is an approximation
}
\seealso{
\code{\link[=rph]{rph()}}

\code{\link[=dph]{dph()}}
}
