% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptdalgorithms-package.r
\docType{package}
\name{ptdalgorithms-package}
\alias{ptdalgorithms}
\alias{ptdalgorithms-package}
\title{ptdalgorithms: Phase-type distribution algorithms}
\description{
Fast, memory efficient graph-based algorithms for phase-type distributions. Includes methods to construct, reward transform, and compute moments of phase-type distributions, rewarded phase-type distributions, and multivariate phase-type distribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TobiasRoikjer/PtDAlgorithms/issues}
  \item Report bugs at \url{https://github.com/TobiasRoikjer/PtDAlgorithms}
}

}
\author{
\strong{Maintainer}: Tobias Røikjer \email{tobiasroikjer@gmail.com} [copyright holder]

}
