% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_sample_stop_vertex}
\alias{random_sample_stop_vertex}
\title{Samples \code{n} stopping times of the phase-type distribution}
\usage{
random_sample_stop_vertex(n, phase_type_graph, time)
}
\arguments{
\item{n}{The number of random samples to draw}

\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{time}{The stopping time}
}
\value{
A vector of numeric values, the indices of the vertices at the stopping time
}
\description{
Samples \code{n} stopping times of the phase-type distribution
}
\seealso{
\code{\link[=rph]{rph()}}
}
