% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdph}
\alias{rdph}
\title{Samples \code{n} outcomes of the discrete phase-type distribution}
\usage{
rdph(n, phase_type_graph, rewards = NULL)
}
\arguments{
\item{n}{The number of random samples to draw}

\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{rewards}{Optional rewards, which should be applied to the discrete phase-type distribution. Must have length equal to \code{\link[=vertices_length]{vertices_length()}}}
}
\value{
A numeric vector of the draws
}
\description{
Allows set.seed as any R function, as it seeds with a random number
generated by R.
}
\seealso{
\code{\link[=rph]{rph()}}

\code{\link[=rmdph]{rmdph()}}
}
