% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmdph}
\alias{rmdph}
\title{Samples one outcome of the multivariate discrete phase-type distribution}
\usage{
rmdph(n, phase_type_graph, rewards)
}
\arguments{
\item{n}{The number of random samples to draw}

\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{rewards}{rewards which should be applied to the discrete phase-type distribution. Must have rows equal to \code{\link[=vertices_length]{vertices_length()}} and columns the dimensions of the multivariate discrete phase-type distribution}
}
\value{
A matrix of numeric values, with \code{rewards} number of rows and \code{n} number of columns
}
\description{
Samples one outcome of the multivariate discrete phase-type distribution
}
\seealso{
\code{\link[=rmph]{rmph()}}

\code{\link[=rdph]{rdph()}}
}
