% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rph}
\alias{rph}
\title{Samples \code{n} outcomes of the phase-type distribution}
\usage{
rph(n, phase_type_graph, rewards = NULL)
}
\arguments{
\item{n}{The number of random samples to draw}

\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{rewards}{Optional rewards, which should be applied to the phase-type distribution. Must have length equal to \code{\link[=vertices_length]{vertices_length()}}}
}
\value{
A vector of numeric values
}
\description{
Samples \code{n} outcomes of the phase-type distribution
}
\seealso{
\code{\link[=rmph]{rmph()}}
}
