% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{starting_vertex}
\alias{starting_vertex}
\title{Returns the special starting vertex of the graph}
\usage{
starting_vertex(phase_type_graph)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\value{
The starting vertex
}
\description{
The starting vertex is always added to the graph
after calling \code{\link[=create_graph]{create_graph()}}, and
always has the first index in \code{\link[=vertex_at]{vertex_at()}}
}
