% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{states}
\alias{states}
\title{Returns a matrix where each row is the state of the vertex at that index}
\usage{
states(phase_type_graph)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\value{
A matrix of size \code{\link[=vertices_length]{vertices_length()}} where the rows match the state of the vertex at that index
}
\description{
Returns a matrix where each row is the state of the vertex at that index
}
\examples{
graph <- ptdalgorithms::create_graph(4)
ptdalgorithms::create_vertex(graph, c(1,2,3,4))
ptdalgorithms::create_vertex(graph, c(4,3,3,3))
ptdalgorithms::states(graph) # => 
# 0 0 0 0
# 1 2 3 4
# 4 3 3 3
}
