% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stop_probability}
\alias{stop_probability}
\title{Computes the probability of the Markov Jump Process of the phase-type distribution standing at each vertex after time}
\usage{
stop_probability(phase_type_graph, time, granularity = 0L)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{time}{The time to wait}

\item{granularity}{Optional. Number of jumps per time unit (approximation)}
}
\value{
A numeric vector of the stop probabilities, for each vertex
}
\description{
Computes the probability of the Markov Jump Process of the phase-type distribution standing at each vertex after time
}
\seealso{
\code{\link[=rph]{rph()}}

\code{\link[=dph]{dph()}}

\code{\link[=pph]{pph()}}
}
