% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vertex_at}
\alias{vertex_at}
\title{Returns a vertex at a particular index.}
\usage{
vertex_at(phase_type_graph, index)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}

\item{index}{The index of the vertex to find}
}
\value{
The vertex at index \code{index} in the graph
}
\description{
This method is much faster than calling \code{ptdalgorithms::vertices()[i]}.
Uses 1-indexing like usual R, so first vertex is 1.
}
