% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vertices}
\alias{vertices}
\title{Obtain a list of all vertices in the graph}
\usage{
vertices(phase_type_graph)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\description{
Returns all vertices that have been added to the
graph from either calling \code{find_or_create_vertex} or
\code{create_vertex}. The first vertex in the list is
\emph{always} the starting vertex \code{\link[=starting_vertex]{starting_vertex()}}.
Importantly, for speed, use \code{\link[=vertices_length]{vertices_length()}} to get the number
of added vertices, and use \code{\link[=vertex_at]{vertex_at()}} to
get a vertex at a particular index.
}
\details{
The list of vertices contains any added vertex, even
if it does not have any in-going / out-going edges.
}
\examples{
graph <- create_graph(4)
vertex_a <- find_or_create_vertex(graph, c(1,2,1,0))
vertex_b <- find_or_create_vertex(graph, c(2,0,1,0))
vertices(graph)[[1]] == starting_vertex(graph)
vertices(graph)[[2]] == vertex_at(graph, 2)
vertices_length(graph) == 3
}
\seealso{
\code{\link[=starting_vertex]{starting_vertex()}}

\code{\link[=vertices_length]{vertices_length()}}

\code{\link[=vertex_at]{vertex_at()}}
}
