% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vertices_length}
\alias{vertices_length}
\title{Returns the number of vertices in the graph}
\usage{
vertices_length(phase_type_graph)
}
\arguments{
\item{phase_type_graph}{A reference to the graph created by \code{\link[=create_graph]{create_graph()}}}
}
\value{
An integer of the number of vertices
}
\description{
This method is much faster than calling \code{length(ptdalgorithms::vertices())}
}
\details{
There will \emph{always} be a starting vertex, so the returned
number is at least 1.
}
